Move CredentialValidationError to core/ui for type-safe error display
