+++ feature/mode/presentation/src/commonMain/kotlin/space/be1ski/vibits/feature/mode/presentation/view/ModeSelectionScreen.kt
@@ -172,7 +172,7 @@ private fun CredentialsSetupDialog(
         )
         state.error?.let { error ->
           Text(
-            text = credentialValidationErrorMessage(error.name),
+            text = credentialValidationErrorMessage(error),
             color = MaterialTheme.colorScheme.error,
             style = MaterialTheme.typography.bodySmall,
           )
