+++ feature/settings/presentation/src/commonMain/kotlin/space/be1ski/vibits/feature/settings/presentation/state/SettingsState.kt
@@ -3,7 +3,7 @@ package space.be1ski.vibits.feature.settings.presentation.state
 import space.be1ski.vibits.core.platform.app.AppDetails
 import space.be1ski.vibits.core.platform.locale.AppLanguage
 import space.be1ski.vibits.core.platform.mode.AppMode
-import space.be1ski.vibits.feature.auth.domain.model.CredentialValidationError
+import space.be1ski.vibits.core.ui.form.CredentialValidationError
 import space.be1ski.vibits.feature.settings.domain.model.AppTheme
 
 data class SettingsState(
