+++ feature/settings/presentation/src/commonMain/kotlin/space/be1ski/vibits/feature/settings/presentation/view/SettingsDialogContent.kt
@@ -166,7 +166,7 @@ private fun SettingsDialogBody(
     )
     state.validationError?.let { error ->
       Text(
-        credentialValidationErrorMessage(error.name),
+        credentialValidationErrorMessage(error),
         color = MaterialTheme.colorScheme.error,
         style = MaterialTheme.typography.bodySmall,
       )
